refactor: extract runWikiLint to lib/lint.ts for MCP sharing
